+++ test/rule/additional/test_additional.py
@@ -1,7 +1,11 @@
+import json
+
 from fastapi.testclient import TestClient
 from sqlmodel import Session
 
 from app.sqlmodels import OrgGroup, Taxon, AdditionalCode, AdditionalRule
+
+from ...mocks import mock_make_search_request
 
 
 class TestAdditional:
@@ -9,7 +13,13 @@ class TestAdditional:
 
     Fixtures for database and authentication come from ../conftest.py"""
 
-    def test_additional(self, client: TestClient):
+    def test_additional(self, client: TestClient, mocker):
+        # Mock the Indicia warehouse.
+        mocker.patch(
+            'app.species.indicia.make_search_request',
+            mock_make_search_request
+        )
+
         # Get database connection from client.
         engine = client.app.context['engine']
         with Session(engine) as db:
@@ -82,6 +92,12 @@ class TestAdditional:
             assert result[0]['code'] == 1
             assert result[0]['text'] == 'Rare'
 
+            # Request additional rules for invlid tvk.
+            response = client.get('/rules/additional/tvk/ABC123')
+            assert response.status_code == 404
+            result = json.loads(response.text)
+            assert result['detail'] == 'TVK ABC123 not recognised.'
+
             # Request additional rules for organism key.
             response = client.get(
                 f'/rules/additional/organism_key/{taxon.organism_key}')
@@ -92,3 +108,9 @@ class TestAdditional:
             assert result[0]['group'] == 'group1'
             assert result[0]['code'] == 1
             assert result[0]['text'] == 'Rare'
+
+            # Request additional rules for invlaid organism key.
+            response = client.get('/rules/additional/organism_key/ABC123')
+            assert response.status_code == 200
+            result = response.json()
+            assert len(result) == 0